requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pycryptodomex==3.17
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.2.0
threadpoolctl==3.1.0
tiktoken==0.2.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
